Initial commit: Exercise
Below is the program code that simulates the purchase of cinema tickets. Two things are verified before purchase:

are there still free seats for a given screening,
whether a given person can purchase a ticket for a given screening. If the movie is only for a certain age group (e.g. 16+), then a person who is not of the required age should not be able to purchase a ticket.
Each subsequent ticket gets an identifier that is a sequential integer starting at 0.

The program code is now completely stored in the main () method. Analyze it and propose division into additional classes and methods. Also eliminate duplicate code.